requirements.txt
deep_translator==1.10.1
bs4==0.0.1
htmldate==1.4.1
pandas==1.5.3
numpy==1.24.2
pendulum==2.1.2
openpyxl==3.0.9
newspaper3k==0.2.8
